#6 [CHORE] admin 폴더명 auth로 변경
