--- src/main/java/com/example/trace/admin/Util/RedisUtil.java
@@ -1,31 +1,0 @@
-package com.example.trace.admin.Util;
-
-import org.springframework.stereotype.Component;
-import org.springframework.data.redis.core.RedisTemplate;
-import java.util.concurrent.TimeUnit;
-@Component
-public class RedisUtil {
-
-    private final RedisTemplate<String, Object> redisTemplate;
-
-    public RedisUtil(RedisTemplate<String, Object> redisTemplate) {
-        this.redisTemplate = redisTemplate;
-    }
-
-
-    public void save(String key, Object val, Long time, TimeUnit timeUnit) {
-        redisTemplate.opsForValue().set(key, val, time, timeUnit);
-    }
-
-    public boolean hasKey(String key) {
-        return Boolean.TRUE.equals(redisTemplate.hasKey(key));
-    }
-
-    public Object get(String key) {
-        return redisTemplate.opsForValue().get(key);
-    }
-
-    public boolean delete(String key) {
-        return Boolean.TRUE.equals(redisTemplate.delete(key));
-    }
-}
