--- src/main/java/com/example/trace/admin/dto/JwtDto.java
@@ -1,7 +1,0 @@
-package com.example.trace.admin.dto;
-
-public record JwtDto(
-        String accessToken,
-        String refreshToken
-) {
-}
