--- src/main/java/com/example/trace/admin/dto/PrincipalDetails.java
@@ -1,64 +1,0 @@
-package com.example.trace.admin.dto;
-
-
-import com.example.trace.admin.domain.User;
-import lombok.Getter;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
-import org.springframework.security.core.userdetails.UserDetails;
-
-import java.util.Collection;
-import java.util.stream.Collectors;
-@Getter
-public class PrincipalDetails implements UserDetails {
-
-    private final User user; // User 객체를 저장
-
-    public PrincipalDetails(User user) {
-        this.user = user;
-    }
-
-
-
-    // 해당 User의 권한을 리턴 하는 곳
-    @Override
-    public Collection<? extends GrantedAuthority> getAuthorities() {
-        return user.getRoleList().stream()
-                .map(SimpleGrantedAuthority::new)
-                .collect(Collectors.toList());
-    }
-
-    @Override
-    public String getPassword() {
-        return user.getPassword();
-    }
-
-    @Override
-    public String getUsername() {
-        return user.getUsername();
-    }
-
-
-    @Override
-    public boolean isAccountNonExpired() {
-        return true;
-    }
-
-    @Override
-    public boolean isAccountNonLocked() {
-        return true;
-    }
-
-    @Override
-    public boolean isCredentialsNonExpired() {
-        return true;
-    }
-
-    @Override
-    public boolean isEnabled() {
-        // 사이트에서 1년 동안 회원이 로그인을 안하면 -> 휴면 계정으로 전환하는 로직이 있다고 치자
-        // user entity의 field에 "Timestamp loginDate"를 하나 만들어주고
-        // (현재 시간 - loginDate) > 1년 -> return false; 로 설정
-        return true;
-    }
-}
