--- src/main/java/com/example/trace/admin/dto/TokenResponse.java
@@ -1,13 +1,0 @@
-package com.example.trace.admin.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-@AllArgsConstructor
-public class TokenResponse {
-    private String accessToken;
-    private String refreshToken;
-}
