--- src/main/java/com/example/trace/admin/models/OIDCDecodePayload.java
@@ -1,21 +1,0 @@
-package com.example.trace.admin.models;
-
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Getter
-@Setter
-@NoArgsConstructor
-public class OIDCDecodePayload {
-    private String iss;
-    private String aud;
-    private String sub;  // user_id in Kakao
-    private long iat;
-    private long auth_time;
-    private long exp;
-    private String nonce;
-    private String nickname;
-    private String picture;
-    private String email;
-}
