--- src/main/java/com/example/trace/admin/models/OIDCPublicKeyResponse.java
@@ -1,14 +1,0 @@
-package com.example.trace.admin.models;
-
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-import java.util.List;
-
-@Getter
-@Setter
-@NoArgsConstructor
-public class OIDCPublicKeyResponse {
-    private List<OIDCPublicKey> keys;
-}
