--- src/main/java/com/example/trace/admin/repository/UserRepository.java
@@ -1,12 +1,0 @@
-package com.example.trace.admin.repository;
-
-import com.example.trace.admin.domain.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-
-@Repository
-public interface UserRepository extends JpaRepository<User, Long> {
-    Optional<User> findByProviderIdAndProvider(String providerId, String provider);
-}
