+++ src/main/java/com/example/trace/auth/dto/DeviceInfo.java
@@ -1,0 +1,15 @@
+package com.example.trace.auth.dto;
+
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@NoArgsConstructor
+public class DeviceInfo {
+    private String deviceId;
+    private String deviceModel;
+    private String osVersion;
+    private String appVersion;
+}
