+++ src/main/java/com/example/trace/auth/dto/JwtDto.java
@@ -1,0 +1,7 @@
+package com.example.trace.auth.dto;
+
+public record JwtDto(
+        String accessToken,
+        String refreshToken
+) {
+}
