+++ src/main/java/com/example/trace/auth/dto/TokenResponse.java
@@ -1,0 +1,13 @@
+package com.example.trace.auth.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+@AllArgsConstructor
+public class TokenResponse {
+    private String accessToken;
+    private String refreshToken;
+}
