+++ src/main/java/com/example/trace/auth/dto/request/KakaoSignupRequest.java
@@ -1,0 +1,25 @@
+
+package com.example.trace.auth.dto.request;
+
+import com.example.trace.auth.dto.DeviceInfo;
+import jakarta.validation.constraints.NotBlank;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@NoArgsConstructor
+public class KakaoSignupRequest {
+    @NotBlank
+    private String idToken;
+
+    // Additional user information for signup
+    private Long ProviderId;
+    private String nickname;
+    private String email;
+    private String profileImage;
+    // Add other necessary fields
+
+    private DeviceInfo deviceInfo;
+}
