+++ src/main/java/com/example/trace/auth/models/OIDCPublicKeyResponse.java
@@ -1,0 +1,14 @@
+package com.example.trace.auth.models;
+
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+import java.util.List;
+
+@Getter
+@Setter
+@NoArgsConstructor
+public class OIDCPublicKeyResponse {
+    private List<OIDCPublicKey> keys;
+}
