+++ src/main/java/com/example/trace/auth/repository/UserRepository.java
@@ -1,0 +1,12 @@
+package com.example.trace.auth.repository;
+
+import com.example.trace.auth.domain.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.Optional;
+
+@Repository
+public interface UserRepository extends JpaRepository<User, Long> {
+    Optional<User> findByProviderIdAndProvider(String providerId, String provider);
+}
